refacto(api): move endpoint into domain files
