+++ crunch/api/client.py
@@ -1,10 +1,95 @@
 import os
+import urllib.parse
 
-from .. import store, constants
+import requests
 
-from .auth import Auth, ApiKeyAuth, NoneAuth
-from .endpoints import EndpointClient
-from .models.competitions import CompetitionCollection
+from .. import constants, store
+from .auth import ApiKeyAuth, Auth, NoneAuth
+from .errors import ApiException
+from .domain.check import CheckEndpointMixin
+from .domain.competition import (CompetitionCollection,
+                                  CompetitionEndpointMixin)
+from .domain.crunch import CrunchEndpointMixin
+from .domain.data_release import DataReleaseEndpointMixin
+from .domain.phase import PhaseEndpointMixin
+from .domain.prediction import PredictionEndpointMixin
+from .domain.project import ProjectEndpointMixin
+from .domain.round import RoundEndpointMixin
+from .domain.score import ScoreEndpointMixin
+
+
+class EndpointClient(
+    requests.Session,
+    CheckEndpointMixin,
+    CompetitionEndpointMixin,
+    CrunchEndpointMixin,
+    DataReleaseEndpointMixin,
+    PhaseEndpointMixin,
+    PredictionEndpointMixin,
+    ProjectEndpointMixin,
+    RoundEndpointMixin,
+    ScoreEndpointMixin,
+):
+
+    def __init__(
+        self,
+        base_url: str,
+        auth: Auth
+    ):
+        super().__init__()
+
+        self.base_url = base_url
+        self.auth_ = auth
+
+    def request(self, method, url, *args, **kwargs):
+        headers = kwargs.pop("headers", {})
+        params = kwargs.pop("params", {})
+        data = kwargs.pop("data", None)
+
+        self.auth_.apply(headers, params, data)
+
+        return super().request(
+            method,
+            urllib.parse.urljoin(self.base_url, url),
+            *args,
+            headers=headers,
+            params=params,
+            data=data,
+            **kwargs,
+        )
+
+    def _raise_for_status(
+        self,
+        response: requests.Response,
+    ):
+        try:
+            response.raise_for_status()
+        except requests.exceptions.HTTPError as error:
+            content = error.response.json()
+
+            code = content.pop("code", "")
+            message = content.pop("message", "")
+
+            raise ApiException(
+                f"{code}: {message}"
+            )
+
+    def _result(
+        self,
+        response: requests.Response,
+        json=False,
+        binary=False
+    ):
+        assert not (json and binary)
+        self._raise_for_status(response)
+
+        if json:
+            return response.json()
+
+        if binary:
+            return response.content
+
+        return response.text
 
 
 class Client:
