+++ crunch/api/domain/crunch.py
@@ -4,8 +4,8 @@ import typing
 import dataclasses_json
 
 from ..identifiers import CrunchIdentifierType
-from .phases import Phase
-from .resource import Collection, Model
+from .phase import Phase
+from ..resource import Collection, Model
 
 
 class Crunch(Model):
@@ -89,3 +89,33 @@ class CrunchCollection(Collection):
             )
             for item in response
         ]
+
+
+class CrunchEndpointMixin:
+
+    def list_crunches(
+        self,
+        competition_identifier,
+        round_identifier,
+        phase_identifier
+    ):
+        return self._result(
+            self.get(
+                f"/v2/competitions/{competition_identifier}/rounds/{round_identifier}/phases/{phase_identifier}/crunches"
+            ),
+            json=True
+        )
+
+    def get_crunch(
+        self,
+        competition_identifier,
+        round_identifier,
+        phase_identifier,
+        crunch_identifier
+    ):
+        return self._result(
+            self.get(
+                f"/v2/competitions/{competition_identifier}/rounds/{round_identifier}/phases/{phase_identifier}/crunches/{crunch_identifier}"
+            ),
+            json=True
+        )
