+++ crunch/api/domain/data_release.py
@@ -5,8 +5,8 @@ import typing
 
 import dataclasses_json
 
-from .competitions import Competition
-from .resource import Collection, Model
+from .competition import Competition
+from ..resource import Collection, Model
 
 
 class DataReleaseTargetResolution(enum.Enum):
@@ -211,3 +211,33 @@ class DataReleaseCollection(Collection):
             )
             for item in response
         ]
+
+
+class DataReleaseEndpointMixin:
+
+    def list_data_releases(
+        self,
+        competition_identifier
+    ):
+        return self._result(
+            self.get(
+                f"/v1/competitions/{competition_identifier}/data-releases"
+            ),
+            json=True
+        )
+
+    def get_data_release(
+        self,
+        competition_identifier,
+        number,
+        include_splits: bool = False
+    ):
+        return self._result(
+            self.get(
+                f"/v1/competitions/{competition_identifier}/data-releases/{number}",
+                params={
+                    "includeSplits": include_splits
+                }
+            ),
+            json=True
+        )
