+++ crunch/api/domain/score.py
@@ -3,8 +3,8 @@ import typing
 
 import dataclasses_json
 
-from .predictions import Prediction
-from .resource import Collection, Model
+from .prediction import Prediction
+from ..resource import Collection, Model
 
 
 @dataclasses_json.dataclass_json(
@@ -68,3 +68,32 @@ class ScoreCollection(Collection):
             )
             for item in response
         ]
+
+
+class ScoreEndpointMixin:
+
+    def list_scores(
+        self,
+        competition_identifier,
+        user_identifier,
+        prediction_id
+    ):
+        return self._result(
+            self.get(
+                f"/v2/competitions/{competition_identifier}/projects/{user_identifier}/predictions/{prediction_id}/scores"
+            ),
+            json=True
+        )
+
+    def get_score_summary(
+        self,
+        competition_identifier,
+        user_identifier,
+        prediction_id
+    ):
+        return self._result(
+            self.get(
+                f"/v2/competitions/{competition_identifier}/projects/{user_identifier}/predictions/{prediction_id}/scores/summary"
+            ),
+            json=True
+        )
