--- crunch/api/models/__init__.py
@@ -1,15 +1,0 @@
-from .checks import Check
-from .competitions import Competition
-from .data_releases import (
-    DataRelease,
-    DataReleaseTargetResolution,
-    DataReleaseSplit,
-    DataReleaseSplitGroup,
-    DataReleaseSplitReduced,
-    DataFiles,
-    OriginalFiles,
-    DataFile,
-)
-from .predictions import Prediction
-from .projects import Project
-from .scores import Score
